requirements.txt
# ── API Framework ────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# ── RAG Core ─────────────────────────────────────
langchain==0.2.6
langchain-community==0.2.6
langchain-google-genai==1.0.7
faiss-cpu==1.8.0
sentence-transformers==3.0.1

# ── PDF Processing ───────────────────────────────
pypdf==4.3.1
unstructured==0.14.6

# ── Voice / STT ──────────────────────────────────
#openai-whisper==20231117
torch==2.3.0
numpy==1.26.4

# ── Translation ──────────────────────────────────
deep-translator==1.11.4
requests==2.31.0
langdetect==1.0.9

# ── Evaluation ───────────────────────────────────
ragas==0.1.14
datasets==2.19.0
pandas==2.2.2

# ── Config & Utils ───────────────────────────────
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0

# ── Frontend ─────────────────────────────────────
streamlit==1.35.0
gtts==2.5.1

# ── Dev Tools ────────────────────────────────────
black==24.4.2
pytest==8.2.0
pytest-asyncio==0.23.7